Tests/LibWeb/Text/expected/abspos-relayout-insert-display-none.txt

inserted-rects=0
sibling-before=40,30,70,25
sibling-after=40,30,70,25
partial-layouts=0 full-layouts=1
PASS
